% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_variability.R
\name{check_variability}
\alias{check_variability}
\title{Check that there are no columns with 0 variability}
\usage{
check_variability(data = NULL)
}
\description{
Check that there are no columns with 0 variability
}
\keyword{internal}
