% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_search.R
\name{group_search}
\alias{group_search}
\title{Find subgroup solutions using final patterns.}
\usage{
group_search(subdata, groupcutoff, yvarnames, interact_exogenous,
  predict_with_interactions, interactnames, interact_exogvars, output,
  grppen = NULL, initial_penalties = NULL)
}
\value{
Returns subgroup membership, similarity matrix, modularity, and number of subgroups.
}
\description{
Find subgroup solutions using final patterns.
}
\keyword{internal}
