% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiLASSO-data.R
\docType{data}
\name{sitNeuro}
\alias{sitNeuro}
\title{Medium Node example with 2 exogenous variables}
\format{A list of data frames with 48 individuals, who each have 910 observations on 11 variables.}
\usage{
sitNeuro
}
\description{
This object contains a list of time series data for forty-eight individuals 
with 910 time points and 11 variables.
}
\keyword{datasets}
