% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_visualization.R
\name{network_visualization}
\alias{network_visualization}
\alias{network_vis}
\title{Network Visualization following Regularized Search}
\usage{
network_vis(output = NULL)
}
\arguments{
\item{output}{Output object from network_reg}
}
\description{
This function utilizes output from network_reg to build network plots.
}
